The shell script for the instance upon launch.
This, in it's current form, does not work. When passing the script into the user field of the AWS instance, the portion of the script that attempts to run a command as the postgres user continues to never be passed, and in turn, the script does not create the database and role. 
Most everything else seems to work, it just needs those commands to run, and it should move forward!